🎨  updating structure of functions per file
