+++ src/ft_movements.c
@@ -6,7 +6,7 @@
 /*   By: maria-ol <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/03 17:00:00 by mona              #+#    #+#             */
-/*   Updated: 2025/10/06 20:19:59 by maria-ol         ###   ########.fr       */
+/*   Updated: 2025/10/07 14:18:03 by maria-ol         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,82 +41,6 @@ void	ft_rotate_a_to_top(t_stack **stack_a, t_stack *target)
 }
 
 /**
- * @brief Rotate stack B to correct position for insertion.
- *
- * @param stack_b Pointer to stack B pointer.
- * @param element Pointer to element to be inserted.
- */
-void	ft_rotate_b_to_pos(t_stack **stack_b, t_stack *element)
-{
-	int	cost;
-	int	rotate_up;
-	int	i;
-
-	if (!*stack_b)
-		return ;
-	cost = ft_calc_rb(*stack_b, element);
-	rotate_up = ft_should_rotate_up_b(*stack_b, element);
-	i = 0;
-	while (i < cost)
-	{
-		if (rotate_up)
-			ft_rb(stack_b);
-		else
-			ft_rrb(stack_b);
-		i++;
-	}
-}
-
-/**
- * @brief Perform optimized rotations using combined operations when possible.
- *
- * @param stack_a Pointer to stack A pointer.
- * @param stack_b Pointer to stack B pointer.
- * @param target Pointer to element to move from A to B.
- */
-static void	ft_optimized_rotations(t_stack **stack_a, t_stack **stack_b,
-				t_stack *target)
-{
-	int	cost_a;
-	int	cost_b;
-	int	rotate_up_a;
-	int	rotate_up_b;
-	int	min_cost;
-
-	cost_a = ft_calc_ra(*stack_a, target);
-	cost_b = ft_calc_rb(*stack_b, target);
-	rotate_up_a = ft_should_rotate_up_a(*stack_a, target);
-	rotate_up_b = ft_should_rotate_up_b(*stack_b, target);
-	if (rotate_up_a == rotate_up_b)
-	{
-		if (cost_a < cost_b)
-			min_cost = cost_a;
-		else
-			min_cost = cost_b;
-		ft_combined_rotations(stack_a, stack_b, rotate_up_a, min_cost);
-		cost_a -= min_cost;
-		cost_b -= min_cost;
-	}
-	ft_individual_rotations_a(stack_a, cost_a, rotate_up_a);
-	ft_individual_rotations_b(stack_b, cost_b, rotate_up_b);
-}
-
-/**
- * @brief Execute the best move to transfer element from A to B.
- *
- * @param stack_a Pointer to stack A pointer.
- * @param stack_b Pointer to stack B pointer.
- * @param target Pointer to element to move from A to B.
- */
-void	ft_best_move(t_stack **stack_a, t_stack **stack_b, t_stack *target)
-{
-	if (!stack_a || !*stack_a || !target)
-		return ;
-	ft_optimized_rotations(stack_a, stack_b, target);
-	ft_pa(stack_a, stack_b);
-}
-
-/**
  * @brief Execute optimal move from B to A with combined rotations.
  *
  * Calculates rotation costs for both stacks, uses combined operations
@@ -146,4 +70,36 @@ void	ft_best_move_b_to_a(t_stack **stack_a, t_stack **stack_b,
 	ft_individual_rotations_a(stack_a, costs[0], costs[2]);
 	ft_individual_rotations_b(stack_b, costs[1], rotate_up_b);
 	ft_pa(stack_a, stack_b);
+}
+
+/**
+ * @brief Find target element in A for insertion from B.
+ *
+ * @param stack_a Stack A.
+ * @param value Value to find target for.
+ * @return Pointer to target element in A.
+ */
+t_stack	*ft_find_target_in_a(t_stack *stack_a, int value)
+{
+	t_stack	*current;
+	t_stack	*best_match;
+	t_stack	*min_element;
+
+	if (!stack_a)
+		return (NULL);
+	current = stack_a;
+	best_match = NULL;
+	min_element = stack_a;
+	while (current)
+	{
+		if (current->index < min_element->index)
+			min_element = current;
+		if (current->index > value
+			&& (!best_match || current->index < best_match->index))
+			best_match = current;
+		current = current->next;
+	}
+	if (!best_match)
+		best_match = min_element;
+	return (best_match);
 }
