+++ src/ft_sorting.c
@@ -6,7 +6,7 @@
 /*   By: maria-ol <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/02 15:00:00 by maria-ol          #+#    #+#             */
-/*   Updated: 2025/10/06 19:35:43 by maria-ol         ###   ########.fr       */
+/*   Updated: 2025/10/07 14:24:26 by maria-ol         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
